Rewrite the Square class to modernize the coding standards.
Unfortunately, Square turns out to be an essential component of the GameBoard
class, which meant rewriting portions of that to allow the program to compile.
The changes were not written with performance in mind, so the program is now
significantly slower. However, it no longer crashes at the end of a game.

Rewriting the entire game board is probably the next step.

Signed-off-by: Jason Lynch <jason@calindora.com>