Implement alpha-beta pruning and fix conversion away from negamax.
The previous commit didn't fix the evaluation function, so the results were
incorrect.

Because of the way cards are referred to in moves, it has become temporarily
impossible to duplicate the board for the player's searching. Because of the way
the board is designed, this won't have any negative effects unless I make the
program multithreaded.

Signed-off-by: Jason Lynch <jason@calindora.com>