Attempt to guess at a better move ordering.
This change doesn't seem to actually help that much, so more research will be
required. There does appear to be a bug with the search algorith, however, as in
one game, a worse scenario emerged than the utility suggested.

The next step is probably evaluation of data structures, as the use of sets
seems to cause more trouble than it's worth (which is essentially nothing, since
the objects are not meaningfully sorted anyway).

Signed-off-by: Jason Lynch <jason@calindora.com>