First round of performance optimizations.
Now that I was more or less happy with the elegance of most of the code, I threw
all caution to the wind and applied a bunch of performance optimizations based
on my profiling. The basic changes so far:

 * As it turns out, std::maps are slow. Instead of mapping squares to cards in a
   map, therefore, I have added a unique ID to each square and instead map using
   a vector. This resulted in perhaps a 30% increase in raw performance, though
   the benchmarks weren't entirely conclusive.
 * I don't know if there's a way to turn off all the locking features in
   std::shared_ptr, but it was proving to be a heavy bottleneck. Creating the
   Moves statically and changing all pointers to them to standard pointers
   resulted in a 50% performance increase, again subject to potentially faulty
   benchmarking.

Note, however, that the code is much less clean and there is additional potential
for memory leaks.

Signed-off-by: Jason Lynch <jason@calindora.com>